move converter to right package
