+++ projects/gloo/pkg/bootstrap/utils.go
@@ -2,13 +2,11 @@ package bootstrap
 
 import (
 	"context"
-	"fmt"
 	"path/filepath"
 	"time"
 
 	"github.com/gogo/protobuf/types"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
-	"github.com/solo-io/solo-kit/pkg/api/v1/resources"
 
 	consulapi "github.com/hashicorp/consul/api"
 	vaultapi "github.com/hashicorp/vault/api"
@@ -17,19 +15,14 @@ import (
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"github.com/solo-io/go-utils/kubeutils"
 	"github.com/solo-io/solo-kit/pkg/api/external/kubernetes/service"
-	skcfgmap "github.com/solo-io/solo-kit/pkg/api/v1/clients/configmap"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/factory"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/crd"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/memory"
 	skkube "github.com/solo-io/solo-kit/pkg/api/v1/resources/common/kubernetes"
 	"github.com/solo-io/solo-kit/pkg/errors"
-	kubev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-
-	skkubeutils "github.com/solo-io/solo-kit/pkg/utils/kubeutils"
-	skprotoutils "github.com/solo-io/solo-kit/pkg/utils/protoutils"
 )
 
 // used for vault and consul key-value storage
@@ -239,7 +232,7 @@ func ArtifactFactoryForSettings(ctx context.Context,
 		return &factory.KubeConfigMapClientFactory{
 			Clientset:       *clientset,
 			Cache:           *kubeCoreCache,
-			CustomConverter: &kubeConverter{},
+			CustomConverter: kubeconverters.NewKubeConfigMapConverter(),
 		}, nil
 	case *v1.Settings_DirectoryArtifactSource:
 		return &factory.FileResourceClientFactory{
@@ -286,62 +279,4 @@ func initializeForKube(ctx context.Context,
 
 	return nil
 
-}
-
-func NewKubeConverter() *kubeConverter {
-	return &kubeConverter{}
-}
-
-type kubeConverter struct{}
-
-func (cc *kubeConverter) FromKubeConfigMap(ctx context.Context, rc *skcfgmap.ResourceClient, configMap *kubev1.ConfigMap) (resources.Resource, error) {
-
-	return cc.FromKubeConfigMapWithResource(ctx, rc.NewResource(), rc.Kind(), configMap)
-}
-
-func (cc *kubeConverter) FromKubeConfigMapWithResource(ctx context.Context, resource resources.Resource, kind string, configMap *kubev1.ConfigMap) (resources.Resource, error) {
-	resourceMap := map[string]interface{}{
-		"data": configMap.Data,
-	}
-
-	if err := skprotoutils.UnmarshalMap(resourceMap, resource); err != nil {
-		return nil, errors.Wrapf(err, "reading configmap data into %v", kind)
-	}
-	resource.SetMetadata(skkubeutils.FromKubeMeta(configMap.ObjectMeta))
-
-	return resource, nil
-}
-
-func (cc *kubeConverter) ToKubeConfigMap(ctx context.Context, rc *skcfgmap.ResourceClient, resource resources.Resource) (*kubev1.ConfigMap, error) {
-	return cc.ToKubeConfigMapSimple(ctx, resource)
-}
-
-func (cc *kubeConverter) ToKubeConfigMapSimple(ctx context.Context, resource resources.Resource) (*kubev1.ConfigMap, error) {
-
-	resourceMap, err := skprotoutils.MarshalMapEmitZeroValues(resource)
-	if err != nil {
-		return nil, errors.Wrapf(err, "marshalling resource as map")
-	}
-	configMapData := make(map[string]string)
-	if dataObj, ok := resourceMap["data"]; ok {
-		if data, ok := dataObj.(map[string]interface{}); ok {
-			for k, v := range data {
-				if stringV, ok := v.(string); ok {
-					configMapData[k] = stringV
-				} else {
-					return nil, fmt.Errorf("resource data value %s of type %T is not a string", k, v)
-				}
-			}
-		} else {
-			return nil, fmt.Errorf("resource data is not map[string]interface{}")
-		}
-	} else {
-		return nil, fmt.Errorf("resource has no data field")
-	}
-
-	meta := skkubeutils.ToKubeMeta(resource.GetMetadata())
-	return &kubev1.ConfigMap{
-		ObjectMeta: meta,
-		Data:       configMapData,
-	}, nil
 }
